refactor: address review — split OOM parser, named MemorySnapshot
Two valid findings from automated review:
- Extract the pure .ips crash-report parsers + JSON helpers out of
  OOMDetector.swift into OOMDetector+CrashReports.swift (separates parsing
  from marker I/O + directory scanning; API unchanged, no test changes).
- Replace GlobalKVCacheBudget's 4-tuple memory-snapshot closure with a named
  MemorySnapshot struct for readability.

Behavior-preserving; 36 affected tests green. Other review findings declined
(see PR discussion): hot-path/blocking-IO flags are per-request not per-step
and caching would make the OOM guard stale; classifier is not duplicated.
